改进立即数的edge case算法，现在不再依赖ub 修复JmpLike指令在操作数接近INT_MIN, INT_MAX时存在一个错误 添加更多寄存器 添加syscall, jle, jne指令 添加所有JmpLike的ret版本（如re表示return if equals） 统一内存架构 改进后优化器

+++ src/cpp/ir/ops/Mul.h
@@ -19,7 +19,7 @@ private:
                                               const std::shared_ptr<Register> &target) {
         if (const auto &immediate = INSTANCEOF_SHARED(value, Immediate)) {
             builder.appendLine(fmt::format("scoreboard players set {} vm_regs {}",
-                                       target->getName(), immediate->getValue()));
+                                       target->getName(), static_cast<i32>(immediate->getValue())));
             return target;
         } else if (const auto &ptr = INSTANCEOF_SHARED(value, Ptr)) {
             builder.appendLine(ptr->loadTo(*target));
